String 타입으로 프론트에서 받아오기/list로 mypost,myrecruit 둘다 보내기/시간 설정

+++ src/main/java/com/week06/team01_week06_project/dto/response/GamePostResDto.java
@@ -16,7 +16,7 @@ public class GamePostResDto {
     private String gameName;
     private String content;
     private List<String> inGameNickname;
-    private int numberOfPeople;
+    private long numberOfPeople;
     private Long numberOdRecruited;
     private boolean recruitStatus;
     private String imgUrl;
